file select moved to window manager, file stuff worked on
for some reason encode was using file instead of self.file. because the encodeAndValue class is now invoked before file is set, a setDefault function is used because everything uses file and file isn't given through __init__ anymore. a getter and setter for file is also now used to the file select menu doesn't have to return anything and file can be set directly in encodeAndValue